Add a network crate
