+++ network/CHANGELOG.md
@@ -1,0 +1,3 @@
+# 0.0.0 - Initial dummy release
+
+- Empty crate to reserve the name on crates.io
