+++ network/README.md
@@ -1,0 +1,7 @@
+# Network
+
+Provides the `Network` type.
+
+## Minimum Supported Rust Version (MSRV)
+
+This library should always compile with any combination of features on **Rust 1.74.0**.
